Fonctionnel
Décrémenter la quantité de produi lors d'une vente

mettre les image en nullable

un bouton creer une vente et un autre valider la vente

+++ solution/MyPopuStore/BU/CategoryPriceServices.cs
@@ -10,16 +10,22 @@ namespace MyPopuStore.BU
     using DB = DAL.DB;
     class CategoryPriceServices
     {
-        public static void add(string color,decimal price) {
-            using (DB.MyPopupStoreDBContext db = new DB.MyPopupStoreDBContext())
+        public static void Add(string color,decimal price) {
+            using (DB.MyPopupStoreDBContext db = new())
             {
                 db.Add(new CategoryPrice() { Color = color, Price = price });
                 db.SaveChanges();
             }
         }
-        public static void delete(CategoryPrice categoryPrice) {
-            using (DB.MyPopupStoreDBContext db = new DB.MyPopupStoreDBContext())
+        public static void Delete(CategoryPrice categoryPrice) {
+            if (CategoryPriceServices.CategoryPriceIsUsed(categoryPrice.CategoryPriceId))
             {
+                throw new Exception("Impossible de supprimer un prix appliqué à au moins un produit.");
+            }
+
+            using (DB.MyPopupStoreDBContext db = new())
+            {
+                
                 db.Remove(categoryPrice);
                 db.SaveChanges();
             }
@@ -27,19 +33,27 @@ namespace MyPopuStore.BU
 
         public static CategoryPrice GetPrice(int ID)
         {
-            CategoryPrice category = null;
             using (MyPopupStoreDBContext db = new())
             {
-                category = db.CategoryPrices.Find(ID);
+                CategoryPrice category = db.CategoryPrices.Find(ID);
+                if (category == null) throw new Exception("Catégorie de prix inexistant");
+                return category;
             }
-            return category==null ? new CategoryPrice() { Color = "#fff", Price = 0 } : category;
         }
 
         public static List<CategoryPrice> GetAllPrice()
         {
-            using (DB.MyPopupStoreDBContext db = new DB.MyPopupStoreDBContext())
+            using (DB.MyPopupStoreDBContext db = new())
             {
                 return db.CategoryPrices.ToList();
+            }
+        }
+
+        public static bool CategoryPriceIsUsed(int catId)
+        {
+            using (DB.MyPopupStoreDBContext db = new())
+            {
+                return db.Products.Where(e => e.CategoryPriceId == catId).Any();
             }
         }
     }
